refactor: 기존 event 패키지 변경, record 로 변경
